Add files via upload
This is the first attempt with finding the linear regression with the CO2 dataset. The problem that keeps occurring is the array xAxis keeps throwing an 'index out of range' error while trying to initialize the 2-D list. I am going to try to use NumPy and make a 2-D array out of that.